chore: data manager 프로토콜 구현체 파일로 이동
